Moving upload certificate to a new component
